api_packing_images/requirements.txt
pydantic
fastapi 
uvicorn
httpx
-e git+https://github.com/ArnoSchiller/PackingUtils.git#egg=packutils
matplotlib
pillow
numpy